Typescript support: first step (#392)
* bot/start,jobs/pending_payments: convert js to ts

Empty tsconfig.json file is added so that we can compile all ts
files using tsc command.


* Add strict to tsconfig.json

* Add `pre` hook for some `npm run` commands

We need to compile TypeScript files before running the code.

* Update dependencies

The typescript package is required. Also we need the latest
version of @types/node, otherwise we get this error we run
the `npm start` & `npm run dev` commands:

```
~\projs\bot ❯ npm start

> lnp2pbot@0.8.1 prestart
> tsc

node_modules/@types/node/globals.d.ts:72:13 - error TS2403: Subsequent variable declarations must have the same type.  Variable 'AbortSignal' must be of type '{ new (): AbortSignal; prototype: AbortSignal; abort(reason?: any): AbortSignal; timeout(milliseconds: number): AbortSignal; }', but here has type '{ new (): AbortSignal; prototype: AbortSignal; }'.

72 declare var AbortSignal: {
               ~~~~~~~~~~~

  node_modules/typescript/lib/lib.dom.d.ts:2304:13
    2304 declare var AbortSignal: {
                     ~~~~~~~~~~~
    'AbortSignal' was also declared here.

Found 1 error in node_modules/@types/node/globals.d.ts:72
```

---------

+++ bot/start.ts
@@ -1,5 +1,9 @@
-const { Telegraf, session } = require('telegraf');
-const { I18n } = require('@grammyjs/i18n');
+import { Telegraf, session, Context } from 'telegraf';
+import { I18n, I18nContext } from '@grammyjs/i18n';
+import { Message } from 'typegram'
+import { UserDocument } from '../models/user'
+import { FilterQuery } from 'mongoose';
+
 const { limit } = require('@grammyjs/ratelimiter');
 const schedule = require('node-schedule');
 const {
@@ -64,42 +68,57 @@ const {
 } = require('../jobs');
 const logger = require('../logger');
 
-const askForConfirmation = async (user, command) => {
+export interface MainContext extends Context {
+  match: Array<string> | null;
+  i18n: I18nContext;
+  user: UserDocument;
+  admin: UserDocument;
+}
+
+interface OrderQuery {
+  status?: string;
+  buyer_id?: string;
+  seller_id?: string;
+}
+
+const askForConfirmation = async (user: UserDocument, command: string) => {
   try {
-    const where = {};
-    if (command == '/cancel') {
-      where.$and = [
-        { $or: [{ buyer_id: user._id }, { seller_id: user._id }] },
-        {
-          $or: [
-            { status: 'ACTIVE' },
-            { status: 'PENDING' },
-            { status: 'FIAT_SENT' },
-            { status: 'DISPUTE' },
-          ],
-        },
-      ];
+    if (command === '/cancel') {
+      const where: FilterQuery<OrderQuery> = {
+        $and: [
+          { $or: [{ buyer_id: user._id }, { seller_id: user._id }] },
+          {
+            $or: [
+              { status: 'ACTIVE' },
+              { status: 'PENDING' },
+              { status: 'FIAT_SENT' },
+              { status: 'DISPUTE' },
+            ],
+          },
+        ]
+      };
       const orders = await Order.find(where);
-
       return orders;
-    } else if (command == '/fiatsent') {
-      where.$and = [{ buyer_id: user._id }, { status: 'ACTIVE' }];
+    } else if (command === '/fiatsent') {
+      const where: FilterQuery<OrderQuery> = {
+        $and: [{ buyer_id: user._id }, { status: 'ACTIVE' }]
+      };
       const orders = await Order.find(where);
-
       return orders;
-    } else if (command == '/release') {
-      where.$and = [
-        { seller_id: user._id },
-        {
-          $or: [
-            { status: 'ACTIVE' },
-            { status: 'FIAT_SENT' },
-            { status: 'DISPUTE' },
-          ],
-        },
-      ];
+    } else if (command === '/release') {
+      const where: FilterQuery<OrderQuery> = {
+        $and: [
+          { seller_id: user._id },
+          {
+            $or: [
+              { status: 'ACTIVE' },
+              { status: 'FIAT_SENT' },
+              { status: 'DISPUTE' },
+            ],
+          },
+        ]
+      };
       const orders = await Order.find(where);
-
       return orders;
     }
 
@@ -109,14 +128,25 @@ const askForConfirmation = async (user, command) => {
   }
 };
 
-const initialize = (botToken, options) => {
+/*
+ctx.update doesn't initially contain message field and it might be
+added to the ctx.update in specific conditions. Therefore we need
+to check the existence of message in ctx.update. ctx.update.message.text
+has the same condition.
+
+The problem mentioned above is similar to this issue:
+https://github.com/telegraf/telegraf/issues/1319#issuecomment-766360594
+*/
+const ctxUpdateAssertMsg = "ctx.update.message.text is not available.";
+
+const initialize = (botToken: string, options: Partial<Telegraf.Options<MainContext>>): Telegraf<MainContext> => {
   const i18n = new I18n({
     defaultLanguageOnMissing: true, // implies allowMissing = true
     directory: 'locales',
     useSession: true,
   });
 
-  const bot = new Telegraf(botToken, options);
+  const bot = new Telegraf<MainContext>(botToken, options);
   bot.catch(err => {
     logger.error(err);
   });
@@ -130,7 +160,7 @@ const initialize = (botToken, options) => {
   // We schedule pending payments job
   schedule.scheduleJob(
     `*/${process.env.PENDING_PAYMENT_WINDOW} * * * *`,
-    async () => {
+    async (): Promise<void> => {
       await attemptPendingPayments(bot);
     }
   );
@@ -147,7 +177,7 @@ const initialize = (botToken, options) => {
     await calculateEarnings();
   });
 
-  schedule.scheduleJob(`*/5 * * * *`, async () => {
+  schedule.scheduleJob(`*/5 * * * *`, async (): Promise<void> => {
     await attemptCommunitiesPendingPayments(bot);
   });
 
@@ -159,8 +189,11 @@ const initialize = (botToken, options) => {
     await nodeInfo(bot);
   });
 
-  bot.start(async ctx => {
+  bot.start(async (ctx: MainContext) => {
     try {
+      if (!('message' in ctx.update) || !('text' in ctx.update.message)){
+        throw new Error(ctxUpdateAssertMsg);
+      }
       const tgUser = ctx.update.message.from;
       if (!tgUser.username) return await messages.nonHandleErrorMessage(ctx);
 
@@ -171,7 +204,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('maintenance', superAdminMiddleware, async ctx => {
+  bot.command('maintenance', superAdminMiddleware, async (ctx: MainContext): Promise<void> => {
     try {
       const [val] = await validateParams(ctx, 2, '\\<_on/off_\\>');
       if (!val) return;
@@ -190,7 +223,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.on('text', userMiddleware, async (ctx, next) => {
+  bot.on('text', userMiddleware, async (ctx: MainContext, next: () => void) => {
     try {
       const config = await Config.findOne({ maintenance: true });
       if (config) {
@@ -203,7 +236,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('version', async ctx => {
+  bot.command('version', async (ctx: MainContext) => {
     try {
       const pckg = require('../package.json');
       await ctx.reply(pckg.version);
@@ -219,8 +252,11 @@ const initialize = (botToken, options) => {
 
   bot.command('release', userMiddleware, async ctx => {
     try {
+      if (!('message' in ctx.update) || !('text' in ctx.update.message)){
+        throw new Error(ctxUpdateAssertMsg);
+      }
       const params = ctx.update.message.text.split(' ');
-      const [command, orderId] = params.filter(el => el);
+      const [command, orderId] = params.filter((el) => el);
 
       if (!orderId) {
         const orders = await askForConfirmation(ctx.user, command);
@@ -239,7 +275,7 @@ const initialize = (botToken, options) => {
 
   DisputeModule.configure(bot);
 
-  bot.command('cancelorder', adminMiddleware, async ctx => {
+  bot.command('cancelorder', adminMiddleware, async (ctx: MainContext) => {
     try {
       const [orderId] = await validateParams(ctx, 2, '\\<_order id_\\>');
 
@@ -305,10 +341,13 @@ const initialize = (botToken, options) => {
 
   // We allow users cancel pending orders,
   // pending orders are the ones that are not taken by another user
-  bot.command('cancel', userMiddleware, async ctx => {
+  bot.command('cancel', userMiddleware, async (ctx: MainContext) => {
     try {
+      if (!('message' in ctx.update) || !('text' in ctx.update.message)){
+        throw new Error(ctxUpdateAssertMsg);
+      }
       const params = ctx.update.message.text.split(' ');
-      const [command, orderId] = params.filter(el => el);
+      const [command, orderId] = params.filter((el) => el);
 
       if (!orderId) {
         const orders = await askForConfirmation(ctx.user, command);
@@ -327,7 +366,7 @@ const initialize = (botToken, options) => {
 
   // We allow users cancel all pending orders,
   // pending orders are the ones that are not taken by another user
-  bot.command('cancelall', userMiddleware, async ctx => {
+  bot.command('cancelall', userMiddleware, async (ctx: MainContext) => {
     try {
       const orders = await ordersActions.getOrders(ctx, ctx.user, 'PENDING');
 
@@ -347,7 +386,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('settleorder', adminMiddleware, async ctx => {
+  bot.command('settleorder', adminMiddleware, async (ctx: MainContext) => {
     try {
       const [orderId] = await validateParams(ctx, 2, '\\<_order id_\\>');
 
@@ -404,7 +443,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('checkorder', superAdminMiddleware, async ctx => {
+  bot.command('checkorder', superAdminMiddleware, async (ctx: MainContext) => {
     try {
       const [orderId] = await validateParams(ctx, 2, '\\<_order id_\\>');
       if (!orderId) return;
@@ -422,7 +461,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('checkinvoice', superAdminMiddleware, async ctx => {
+  bot.command('checkinvoice', superAdminMiddleware, async (ctx: MainContext) => {
     try {
       const [orderId] = await validateParams(ctx, 2, '\\<_order id_\\>');
       if (!orderId) return;
@@ -445,7 +484,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('resubscribe', superAdminMiddleware, async ctx => {
+  bot.command('resubscribe', superAdminMiddleware, async (ctx: MainContext) => {
     try {
       const [hash] = await validateParams(ctx, 2, '\\<_hash_\\>');
 
@@ -456,12 +495,12 @@ const initialize = (botToken, options) => {
       if (!order) return;
       await subscribeInvoice(bot, hash, true);
       ctx.reply(`hash resubscribed`);
-    } catch (error) {
+    } catch (error: any) {
       logger.error(`/resubscribe command error: ${error.toString()}`);
     }
   });
 
-  bot.command('help', userMiddleware, async ctx => {
+  bot.command('help', userMiddleware, async (ctx: MainContext) => {
     try {
       await messages.helpMessage(ctx);
     } catch (error) {
@@ -470,10 +509,13 @@ const initialize = (botToken, options) => {
   });
 
   // Only buyers can use this command
-  bot.command('fiatsent', userMiddleware, async ctx => {
+  bot.command('fiatsent', userMiddleware, async (ctx: MainContext) => {
     try {
+      if (!('message' in ctx.update) || !('text' in ctx.update.message)){
+        throw new Error(ctxUpdateAssertMsg);
+      }
       const params = ctx.update.message.text.split(' ');
-      const [command, orderId] = params.filter(el => el);
+      const [command, orderId] = params.filter((el) => el);
 
       if (!orderId) {
         const orders = await askForConfirmation(ctx.user, command);
@@ -490,7 +532,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('ban', adminMiddleware, async ctx => {
+  bot.command('ban', adminMiddleware, async (ctx: MainContext) => {
     try {
       let [username] = await validateParams(
         ctx,
@@ -533,7 +575,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('unban', adminMiddleware, async ctx => {
+  bot.command('unban', adminMiddleware, async (ctx: MainContext) => {
     try {
       let [username] = await validateParams(
         ctx,
@@ -559,7 +601,7 @@ const initialize = (botToken, options) => {
             ctx.admin.default_community_id
           );
           community.banned_users = community.banned_users.filter(
-            el => el.id !== user.id
+            (el: any) => el.id !== user.id
           );
           await community.save();
         } else {
@@ -575,7 +617,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('setaddress', userMiddleware, async ctx => {
+  bot.command('setaddress', userMiddleware, async (ctx: MainContext) => {
     try {
       const [lightningAddress] = await validateParams(
         ctx,
@@ -602,7 +644,7 @@ const initialize = (botToken, options) => {
   });
 
   // Only buyers can use this command
-  bot.command('setinvoice', userMiddleware, async ctx => {
+  bot.command('setinvoice', userMiddleware, async (ctx: MainContext) => {
     try {
       const [orderId, lnInvoice] = await validateParams(
         ctx,
@@ -680,46 +722,61 @@ const initialize = (botToken, options) => {
 
   OrdersModule.configure(bot);
 
-  bot.action('addInvoiceBtn', userMiddleware, async ctx => {
+  bot.action('addInvoiceBtn', userMiddleware, async (ctx: MainContext) => {
     await addInvoice(ctx, bot);
   });
 
-  bot.action('cancelAddInvoiceBtn', userMiddleware, async ctx => {
+  bot.action('cancelAddInvoiceBtn', userMiddleware, async (ctx: MainContext) => {
     await cancelAddInvoice(ctx);
   });
 
-  bot.action('showHoldInvoiceBtn', userMiddleware, async ctx => {
+  bot.action('showHoldInvoiceBtn', userMiddleware, async (ctx: MainContext) => {
     await showHoldInvoice(ctx, bot);
   });
 
-  bot.action('cancelShowHoldInvoiceBtn', userMiddleware, async ctx => {
+  bot.action('cancelShowHoldInvoiceBtn', userMiddleware, async (ctx: MainContext) => {
     await cancelShowHoldInvoice(ctx);
   });
 
-  bot.action(/^showStarBtn\(([1-5]),(\w{24})\)$/, userMiddleware, async ctx => {
+  bot.action(/^showStarBtn\(([1-5]),(\w{24})\)$/, userMiddleware, async (ctx: MainContext) => {
+    if (ctx.match === null) {
+      throw new Error("ctx.match should not be null");
+    }
     await rateUser(ctx, bot, ctx.match[1], ctx.match[2]);
   });
 
-  bot.action(/^addInvoicePHIBtn_([0-9a-f]{24})$/, userMiddleware, async ctx => {
+  bot.action(/^addInvoicePHIBtn_([0-9a-f]{24})$/, userMiddleware, async (ctx: MainContext) => {
+    if (ctx.match === null) {
+      throw new Error("ctx.match should not be null");
+    }
     await addInvoicePHI(ctx, bot, ctx.match[1]);
   });
 
-  bot.action(/^cancel_([0-9a-f]{24})$/, userMiddleware, async ctx => {
+  bot.action(/^cancel_([0-9a-f]{24})$/, userMiddleware, async (ctx: MainContext) => {
+    if (ctx.match === null) {
+      throw new Error("ctx.match should not be null");
+    }
     ctx.deleteMessage();
     await cancelOrder(ctx, ctx.match[1]);
   });
 
-  bot.action(/^fiatsent_([0-9a-f]{24})$/, userMiddleware, async ctx => {
+  bot.action(/^fiatsent_([0-9a-f]{24})$/, userMiddleware, async (ctx: MainContext) => {
+    if (ctx.match === null) {
+      throw new Error("ctx.match should not be null");
+    }
     ctx.deleteMessage();
     await fiatSent(ctx, ctx.match[1]);
   });
 
-  bot.action(/^release_([0-9a-f]{24})$/, userMiddleware, async ctx => {
+  bot.action(/^release_([0-9a-f]{24})$/, userMiddleware, async (ctx: MainContext) => {
+    if (ctx.match === null) {
+      throw new Error("ctx.match should not be null");
+    }
     ctx.deleteMessage();
     await release(ctx, ctx.match[1]);
   });
 
-  bot.command('paytobuyer', superAdminMiddleware, async ctx => {
+  bot.command('paytobuyer', superAdminMiddleware, async (ctx: MainContext) => {
     try {
       const [orderId] = await validateParams(ctx, 2, '\\<_order id_\\>');
       if (!orderId) return;
@@ -743,7 +800,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('listcurrencies', userMiddleware, async ctx => {
+  bot.command('listcurrencies', userMiddleware, async (ctx: MainContext) => {
     try {
       const currencies = getCurrenciesWithPrice();
 
@@ -753,7 +810,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('info', userMiddleware, async ctx => {
+  bot.command('info', userMiddleware, async (ctx: MainContext) => {
     try {
       const config = await Config.findOne({});
       await messages.showInfoMessage(ctx, ctx.user, config);
@@ -762,7 +819,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('showusername', userMiddleware, async ctx => {
+  bot.command('showusername', userMiddleware, async (ctx: MainContext) => {
     try {
       let [show] = await validateParams(ctx, 2, '_yes/no_');
       if (!show) return;
@@ -775,7 +832,7 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('showvolume', userMiddleware, async ctx => {
+  bot.command('showvolume', userMiddleware, async (ctx: MainContext) => {
     try {
       let [show] = await validateParams(ctx, 2, '_yes/no_');
       if (!show) return;
@@ -788,9 +845,9 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.command('exit', userMiddleware, async ctx => {
+  bot.command('exit', userMiddleware, async (ctx: MainContext) => {
     try {
-      if (ctx.message.chat.type !== 'private') return;
+      if (ctx.message?.chat.type !== 'private') return;
 
       await ctx.reply(ctx.i18n.t('not_wizard'));
     } catch (error) {
@@ -798,11 +855,11 @@ const initialize = (botToken, options) => {
     }
   });
 
-  bot.on('text', userMiddleware, async ctx => {
+  bot.on('text', userMiddleware, async (ctx: MainContext) => {
     try {
-      if (ctx.message.chat.type !== 'private') return;
+      if (ctx.message?.chat.type !== 'private') return;
 
-      const text = ctx.message.text;
+      const text = (ctx.message as Message.TextMessage).text;
       let message;
       // If the user is trying to enter a command with first letter uppercase
       if (text[0] === '/' && text[1] === text[1].toUpperCase()) {
@@ -819,7 +876,7 @@ const initialize = (botToken, options) => {
   return bot;
 };
 
-const start = (botToken, options) => {
+const start = (botToken: string, options: Partial<Telegraf.Options<MainContext>>): Telegraf<MainContext> => {
   const bot = initialize(botToken, options);
 
   bot.launch();
@@ -833,4 +890,4 @@ const start = (botToken, options) => {
   return bot;
 };
 
-module.exports = { initialize, start };
+export { initialize, start };
